ecomm/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.4
braintree==4.19.0
cachetools==4.0.0
certifi==2023.5.7
chardet==5.1.0
Click==7.0
colorama==0.4.3
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
dj-database-url==2.0.0
Django==4.2
django-bootstrap4==23.1
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
entrypoints==0.3
exceptiongroup==1.1.1
fasteners==0.14.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==20.1.0
h11==0.14.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
mysqlclient==2.1.1
netifaces==0.10.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
outcome==1.2.0
packaging==20.3
paramiko==2.6.0
pbr==5.11.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.7.3
python-dotenv==1.0.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
selenium==4.9.1
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==1.3.12
sqlparse>=0.3.1
stripe==5.4.0
systemd-python==234
testresources==2.0.1
trio==0.22.0
trio-websocket==0.10.2
ttystatus==0.38
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==2.0.2
usb-creator==0.3.7
wadllib==1.3.3
<<<<<<< HEAD
=======
wsproto==1.2.0
>>>>>>> 26c6feb96d02fe17a07b63c3ddd3bba3789e9778
xkit==0.0.0